Add Listing draft/published validation
