+++ backend/core/package.json
@@ -35,12 +35,12 @@
     "@anchan828/nest-sendgrid": "^0.3.25",
     "@google-cloud/translate": "^6.2.6",
     "@nestjs/cli": "^7.5.1",
-    "@nestjs/common": "^7.4.4",
+    "@nestjs/common": "^7.6.18",
     "@nestjs/config": "^0.5.0",
-    "@nestjs/core": "^7.4.4",
+    "@nestjs/core": "^7.6.18",
     "@nestjs/jwt": "^7.1.0",
     "@nestjs/passport": "^7.1.0",
-    "@nestjs/platform-express": "^7.4.4",
+    "@nestjs/platform-express": "^7.6.18",
     "@nestjs/swagger": "4.7.3",
     "@nestjs/throttler": "^1.1.2",
     "@nestjs/typeorm": "^7.1.0",
