+++ backend/core/src/listings/dto/listing.dto.ts
@@ -1,50 +1,21 @@
 import { Listing } from "../entities/listing.entity"
 import { Expose, plainToClass, Transform, Type } from "class-transformer"
-import {
-  ArrayMaxSize,
-  IsDate,
-  IsDefined,
-  IsNumber,
-  IsOptional,
-  IsString,
-  IsUUID,
-  ValidateNested,
-  IsNumberString,
-  IsEnum,
-  IsArray,
-} from "class-validator"
+import { IsDefined, IsNumber, IsOptional, IsString, ValidateNested } from "class-validator"
 import moment from "moment"
-import {
-  PreferenceCreateDto,
-  PreferenceDto,
-  PreferenceUpdateDto,
-} from "../../preferences/dto/preference.dto"
-import { ApiProperty, OmitType, getSchemaPath } from "@nestjs/swagger"
-import { IdDto } from "../../shared/dto/id.dto"
-import { AddressCreateDto, AddressDto, AddressUpdateDto } from "../../shared/dto/address.dto"
+import { PreferenceDto } from "../../preferences/dto/preference.dto"
+import { OmitType } from "@nestjs/swagger"
+import { AddressDto } from "../../shared/dto/address.dto"
 import { ValidationsGroupsEnum } from "../../shared/types/validations-groups-enum"
 import { ListingStatus } from "../types/listing-status-enum"
-import { ListingFilterKeys } from "../types/listing-filter-keys-enum"
-import { PaginationFactory, PaginationAllowsAllQueryParams } from "../../shared/dto/pagination.dto"
-import { BaseFilter } from "../../shared/dto/filter.dto"
-import { UnitCreateDto, UnitDto, UnitUpdateDto } from "../../units/dto/unit.dto"
+import { UnitDto } from "../../units/dto/unit.dto"
 import { ReservedCommunityTypeDto } from "../../reserved-community-type/dto/reserved-community-type.dto"
-import { AssetCreateDto, AssetDto, AssetUpdateDto } from "../../assets/dto/asset.dto"
-import {
-  ApplicationMethodCreateDto,
-  ApplicationMethodDto,
-  ApplicationMethodUpdateDto,
-} from "../../application-methods/dto/application-method.dto"
-import { ListingEventCreateDto, ListingEventDto, ListingEventUpdateDto } from "./listing-event.dto"
+import { AssetDto } from "../../assets/dto/asset.dto"
+import { ListingEventDto } from "./listing-event.dto"
 import { listingUrlSlug } from "../../shared/url-helper"
-import {
-  UnitsSummaryCreateDto,
-  UnitsSummaryDto,
-  UnitsSummaryUpdateDto,
-} from "../../units-summary/dto/units-summary.dto"
-import { OrderByFieldsEnum } from "../types/listing-orderby-enum"
 import { IdNameDto } from "../../shared/dto/idName.dto"
 import { UserBasicDto } from "../../auth/dto/user-basic.dto"
+import { ApplicationMethodDto } from "../../application-methods/dto/application-method.dto"
+import { UnitsSummaryDto } from "../../units-summary/dto/units-summary.dto"
 
 export class ListingDto extends OmitType(Listing, [
   "applicationAddress",
@@ -351,551 +322,3 @@ export class ListingDto extends OmitType(Listing, [
   countyCode?: string
 }
 
-export class PaginatedListingDto extends PaginationFactory<ListingDto>(ListingDto) {}
-
-export class ListingCreateDto extends OmitType(ListingDto, [
-  "id",
-  "createdAt",
-  "updatedAt",
-  "applicationMethods",
-  "preferences",
-  "events",
-  "image",
-  "leasingAgentAddress",
-  "leasingAgents",
-  "urlSlug",
-  "showWaitlist",
-  "units",
-  "accessibility",
-  "amenities",
-  "buildingAddress",
-  "buildingTotalUnits",
-  "developer",
-  "householdSizeMax",
-  "householdSizeMin",
-  "neighborhood",
-  "petPolicy",
-  "smokingPolicy",
-  "unitsAvailable",
-  "unitAmenities",
-  "servicesOffered",
-  "yearBuilt",
-  "unitsSummarized",
-  "jurisdiction",
-  "reservedCommunityType",
-  "result",
-  "unitsSummary",
-] as const) {
-  @Expose()
-  @IsDefined({ groups: [ValidationsGroupsEnum.default] })
-  @ValidateNested({ groups: [ValidationsGroupsEnum.default], each: true })
-  @Type(() => ApplicationMethodCreateDto)
-  applicationMethods: ApplicationMethodCreateDto[]
-
-  @Expose()
-  @IsDefined({ groups: [ValidationsGroupsEnum.default] })
-  @ValidateNested({ groups: [ValidationsGroupsEnum.default], each: true })
-  @Type(() => PreferenceCreateDto)
-  preferences: PreferenceCreateDto[]
-
-  @Expose()
-  @IsOptional({ groups: [ValidationsGroupsEnum.default] })
-  @ValidateNested({ groups: [ValidationsGroupsEnum.default] })
-  @Type(() => AddressCreateDto)
-  applicationAddress?: AddressCreateDto | null
-
-  @Expose()
-  @IsOptional({ groups: [ValidationsGroupsEnum.default] })
-  @ValidateNested({ groups: [ValidationsGroupsEnum.default] })
-  @Type(() => AddressCreateDto)
-  applicationPickUpAddress?: AddressCreateDto | null
-
-  @Expose()
-  @IsOptional({ groups: [ValidationsGroupsEnum.default] })
-  @ValidateNested({ groups: [ValidationsGroupsEnum.default] })
-  @Type(() => AddressCreateDto)
-  applicationDropOffAddress: AddressCreateDto | null
-
-  @Expose()
-  @IsOptional({ groups: [ValidationsGroupsEnum.default] })
-  @ValidateNested({ groups: [ValidationsGroupsEnum.default] })
-  @Type(() => AddressCreateDto)
-  applicationMailingAddress: AddressCreateDto | null
-
-  @Expose()
-  @IsDefined({ groups: [ValidationsGroupsEnum.default] })
-  @ValidateNested({ groups: [ValidationsGroupsEnum.default], each: true })
-  @Type(() => ListingEventCreateDto)
-  events: ListingEventCreateDto[]
-
-  @Expose()
-  @IsOptional({ groups: [ValidationsGroupsEnum.default] })
-  @ValidateNested({ groups: [ValidationsGroupsEnum.default] })
-  @Type(() => AssetCreateDto)
-  image?: AssetCreateDto | null
-
-  @Expose()
-  @IsOptional({ groups: [ValidationsGroupsEnum.default] })
-  @ValidateNested({ groups: [ValidationsGroupsEnum.default] })
-  @Type(() => AddressCreateDto)
-  leasingAgentAddress?: AddressCreateDto | null
-
-  @Expose()
-  @IsOptional({ groups: [ValidationsGroupsEnum.default] })
-  @IsDefined({ groups: [ValidationsGroupsEnum.default] })
-  @ValidateNested({ groups: [ValidationsGroupsEnum.default], each: true })
-  @Type(() => IdDto)
-  leasingAgents?: IdDto[] | null
-
-  @Expose()
-  @IsDefined({ groups: [ValidationsGroupsEnum.default] })
-  @ValidateNested({ groups: [ValidationsGroupsEnum.default], each: true })
-  @ArrayMaxSize(256, { groups: [ValidationsGroupsEnum.default] })
-  @Type(() => UnitCreateDto)
-  units: UnitCreateDto[]
-
-  @Expose()
-  @IsOptional({ groups: [ValidationsGroupsEnum.default] })
-  @IsString({ groups: [ValidationsGroupsEnum.default] })
-  accessibility?: string | null
-
-  @Expose()
-  @IsOptional({ groups: [ValidationsGroupsEnum.default] })
-  @IsString({ groups: [ValidationsGroupsEnum.default] })
-  amenities?: string | null
-
-  @Expose()
-  @IsDefined({ groups: [ValidationsGroupsEnum.default] })
-  @ValidateNested({ groups: [ValidationsGroupsEnum.default] })
-  @Type(() => AddressCreateDto)
-  buildingAddress: AddressCreateDto
-
-  @Expose()
-  @IsOptional({ groups: [ValidationsGroupsEnum.default] })
-  @IsNumber({}, { groups: [ValidationsGroupsEnum.default] })
-  buildingTotalUnits?: number | null
-
-  @Expose()
-  @IsOptional({ groups: [ValidationsGroupsEnum.default] })
-  @IsString({ groups: [ValidationsGroupsEnum.default] })
-  developer?: string | null
-
-  @Expose()
-  @IsOptional({ groups: [ValidationsGroupsEnum.default] })
-  @IsNumber({}, { groups: [ValidationsGroupsEnum.default] })
-  householdSizeMax?: number | null
-
-  @Expose()
-  @IsOptional({ groups: [ValidationsGroupsEnum.default] })
-  @IsNumber({}, { groups: [ValidationsGroupsEnum.default] })
-  householdSizeMin?: number | null
-
-  @Expose()
-  @IsOptional({ groups: [ValidationsGroupsEnum.default] })
-  @IsString({ groups: [ValidationsGroupsEnum.default] })
-  neighborhood?: string | null
-
-  @Expose()
-  @IsOptional({ groups: [ValidationsGroupsEnum.default] })
-  @IsString({ groups: [ValidationsGroupsEnum.default] })
-  petPolicy?: string | null
-
-  @Expose()
-  @IsOptional({ groups: [ValidationsGroupsEnum.default] })
-  @IsString({ groups: [ValidationsGroupsEnum.default] })
-  smokingPolicy?: string | null
-
-  @Expose()
-  @IsOptional({ groups: [ValidationsGroupsEnum.default] })
-  @IsNumber({}, { groups: [ValidationsGroupsEnum.default] })
-  unitsAvailable?: number | null
-
-  @Expose()
-  @IsOptional({ groups: [ValidationsGroupsEnum.default] })
-  @IsString({ groups: [ValidationsGroupsEnum.default] })
-  unitAmenities?: string | null
-
-  @Expose()
-  @IsOptional({ groups: [ValidationsGroupsEnum.default] })
-  @IsString({ groups: [ValidationsGroupsEnum.default] })
-  servicesOffered?: string | null
-
-  @Expose()
-  @IsOptional({ groups: [ValidationsGroupsEnum.default] })
-  @IsNumber({}, { groups: [ValidationsGroupsEnum.default] })
-  yearBuilt?: number | null
-
-  @Expose()
-  @IsDefined({ groups: [ValidationsGroupsEnum.default] })
-  @ValidateNested({ groups: [ValidationsGroupsEnum.default] })
-  @Type(() => IdDto)
-  jurisdiction: IdDto
-
-  @Expose()
-  @IsOptional({ groups: [ValidationsGroupsEnum.default] })
-  @IsDefined({ groups: [ValidationsGroupsEnum.default] })
-  @ValidateNested({ groups: [ValidationsGroupsEnum.default] })
-  @Type(() => IdDto)
-  reservedCommunityType?: IdDto
-
-  @Expose()
-  @IsOptional({ groups: [ValidationsGroupsEnum.default] })
-  @ValidateNested({ groups: [ValidationsGroupsEnum.default] })
-  @Type(() => AssetCreateDto)
-  result?: AssetCreateDto | null
-
-  @Expose()
-  @IsOptional({ groups: [ValidationsGroupsEnum.default] })
-  @ValidateNested({ groups: [ValidationsGroupsEnum.default], each: true })
-  @Type(() => UnitsSummaryCreateDto)
-  unitsSummary?: UnitsSummaryCreateDto[]
-}
-
-export class ListingUpdateDto extends OmitType(ListingDto, [
-  "id",
-  "createdAt",
-  "updatedAt",
-  "applicationMethods",
-  "preferences",
-  "image",
-  "events",
-  "leasingAgentAddress",
-  "urlSlug",
-  "leasingAgents",
-  "showWaitlist",
-  "units",
-  "accessibility",
-  "amenities",
-  "buildingAddress",
-  "buildingTotalUnits",
-  "developer",
-  "householdSizeMax",
-  "householdSizeMin",
-  "neighborhood",
-  "petPolicy",
-  "smokingPolicy",
-  "unitsAvailable",
-  "unitAmenities",
-  "servicesOffered",
-  "yearBuilt",
-  "unitsSummarized",
-  "jurisdiction",
-  "reservedCommunityType",
-  "result",
-  "unitsSummary",
-] as const) {
-  @Expose()
-  @IsOptional({ groups: [ValidationsGroupsEnum.default] })
-  @IsUUID(4, { groups: [ValidationsGroupsEnum.default] })
-  id?: string
-
-  @Expose()
-  @IsOptional({ groups: [ValidationsGroupsEnum.default] })
-  @IsDate({ groups: [ValidationsGroupsEnum.default] })
-  @Type(() => Date)
-  createdAt?: Date
-
-  @Expose()
-  @IsOptional({ groups: [ValidationsGroupsEnum.default] })
-  @IsDate({ groups: [ValidationsGroupsEnum.default] })
-  @Type(() => Date)
-  updatedAt?: Date
-
-  @Expose()
-  @IsDefined({ groups: [ValidationsGroupsEnum.default] })
-  @ValidateNested({ groups: [ValidationsGroupsEnum.default], each: true })
-  @Type(() => ApplicationMethodUpdateDto)
-  applicationMethods: ApplicationMethodUpdateDto[]
-
-  @Expose()
-  @IsDefined({ groups: [ValidationsGroupsEnum.default] })
-  @ValidateNested({ groups: [ValidationsGroupsEnum.default], each: true })
-  @Type(() => PreferenceUpdateDto)
-  preferences: PreferenceUpdateDto[]
-
-  @Expose()
-  @IsOptional({ groups: [ValidationsGroupsEnum.default] })
-  @ValidateNested({ groups: [ValidationsGroupsEnum.default] })
-  @Type(() => AddressUpdateDto)
-  applicationAddress?: AddressUpdateDto | null
-
-  @Expose()
-  @IsOptional({ groups: [ValidationsGroupsEnum.default] })
-  @ValidateNested({ groups: [ValidationsGroupsEnum.default] })
-  @Type(() => AddressUpdateDto)
-  applicationPickUpAddress?: AddressUpdateDto | null
-
-  @Expose()
-  @IsOptional({ groups: [ValidationsGroupsEnum.default] })
-  @ValidateNested({ groups: [ValidationsGroupsEnum.default] })
-  @Type(() => AddressUpdateDto)
-  applicationDropOffAddress: AddressUpdateDto | null
-
-  @Expose()
-  @IsOptional({ groups: [ValidationsGroupsEnum.default] })
-  @ValidateNested({ groups: [ValidationsGroupsEnum.default] })
-  @Type(() => AddressUpdateDto)
-  applicationMailingAddress: AddressUpdateDto | null
-
-  @Expose()
-  @IsDefined({ groups: [ValidationsGroupsEnum.default] })
-  @ValidateNested({ groups: [ValidationsGroupsEnum.default], each: true })
-  @Type(() => ListingEventUpdateDto)
-  events: ListingEventUpdateDto[]
-
-  @Expose()
-  @IsOptional({ groups: [ValidationsGroupsEnum.default] })
-  @ValidateNested({ groups: [ValidationsGroupsEnum.default] })
-  @Type(() => AssetUpdateDto)
-  image?: AssetUpdateDto
-
-  @Expose()
-  @IsOptional({ groups: [ValidationsGroupsEnum.default] })
-  @ValidateNested({ groups: [ValidationsGroupsEnum.default] })
-  @Type(() => AddressUpdateDto)
-  leasingAgentAddress?: AddressUpdateDto | null
-
-  @Expose()
-  @IsOptional({ groups: [ValidationsGroupsEnum.default] })
-  @IsDefined({ groups: [ValidationsGroupsEnum.default] })
-  @ValidateNested({ groups: [ValidationsGroupsEnum.default], each: true })
-  @Type(() => IdDto)
-  leasingAgents?: IdDto[] | null
-
-  @Expose()
-  @IsDefined({ groups: [ValidationsGroupsEnum.default] })
-  @ValidateNested({ groups: [ValidationsGroupsEnum.default], each: true })
-  @ArrayMaxSize(256, { groups: [ValidationsGroupsEnum.default] })
-  @Type(() => UnitUpdateDto)
-  units: UnitUpdateDto[]
-
-  @Expose()
-  @IsOptional({ groups: [ValidationsGroupsEnum.default] })
-  @IsString({ groups: [ValidationsGroupsEnum.default] })
-  accessibility?: string | null
-
-  @Expose()
-  @IsOptional({ groups: [ValidationsGroupsEnum.default] })
-  @IsString({ groups: [ValidationsGroupsEnum.default] })
-  amenities?: string | null
-
-  @Expose()
-  @IsDefined({ groups: [ValidationsGroupsEnum.default] })
-  @ValidateNested({ groups: [ValidationsGroupsEnum.default] })
-  @Type(() => AddressUpdateDto)
-  buildingAddress: AddressUpdateDto
-
-  @Expose()
-  @IsOptional({ groups: [ValidationsGroupsEnum.default] })
-  @IsNumber({}, { groups: [ValidationsGroupsEnum.default] })
-  buildingTotalUnits?: number | null
-
-  @Expose()
-  @IsOptional({ groups: [ValidationsGroupsEnum.default] })
-  @IsString({ groups: [ValidationsGroupsEnum.default] })
-  developer?: string | null
-
-  @Expose()
-  @IsOptional({ groups: [ValidationsGroupsEnum.default] })
-  @IsNumber({}, { groups: [ValidationsGroupsEnum.default] })
-  householdSizeMax?: number | null
-
-  @Expose()
-  @IsOptional({ groups: [ValidationsGroupsEnum.default] })
-  @IsNumber({}, { groups: [ValidationsGroupsEnum.default] })
-  householdSizeMin?: number | null
-
-  @Expose()
-  @IsOptional({ groups: [ValidationsGroupsEnum.default] })
-  @IsString({ groups: [ValidationsGroupsEnum.default] })
-  neighborhood?: string | null
-
-  @Expose()
-  @IsOptional({ groups: [ValidationsGroupsEnum.default] })
-  @IsString({ groups: [ValidationsGroupsEnum.default] })
-  petPolicy?: string | null
-
-  @Expose()
-  @IsOptional({ groups: [ValidationsGroupsEnum.default] })
-  @IsString({ groups: [ValidationsGroupsEnum.default] })
-  smokingPolicy?: string | null
-
-  @Expose()
-  @IsOptional({ groups: [ValidationsGroupsEnum.default] })
-  @IsNumber({}, { groups: [ValidationsGroupsEnum.default] })
-  unitsAvailable?: number | null
-
-  @Expose()
-  @IsOptional({ groups: [ValidationsGroupsEnum.default] })
-  @IsString({ groups: [ValidationsGroupsEnum.default] })
-  unitAmenities?: string | null
-
-  @Expose()
-  @IsOptional({ groups: [ValidationsGroupsEnum.default] })
-  @IsString({ groups: [ValidationsGroupsEnum.default] })
-  servicesOffered?: string | null
-
-  @Expose()
-  @IsOptional({ groups: [ValidationsGroupsEnum.default] })
-  @IsNumber({}, { groups: [ValidationsGroupsEnum.default] })
-  yearBuilt?: number | null
-
-  @Expose()
-  @IsDefined({ groups: [ValidationsGroupsEnum.default] })
-  @ValidateNested({ groups: [ValidationsGroupsEnum.default] })
-  @Type(() => IdDto)
-  jurisdiction: IdDto
-
-  @Expose()
-  @IsOptional({ groups: [ValidationsGroupsEnum.default] })
-  @IsDefined({ groups: [ValidationsGroupsEnum.default] })
-  @ValidateNested({ groups: [ValidationsGroupsEnum.default] })
-  @Type(() => IdDto)
-  reservedCommunityType?: IdDto
-
-  @Expose()
-  @IsOptional({ groups: [ValidationsGroupsEnum.default] })
-  @ValidateNested({ groups: [ValidationsGroupsEnum.default] })
-  @Type(() => AssetUpdateDto)
-  result?: AssetUpdateDto
-
-  @Expose()
-  @IsOptional({ groups: [ValidationsGroupsEnum.default] })
-  @ValidateNested({ groups: [ValidationsGroupsEnum.default], each: true })
-  @Type(() => UnitsSummaryUpdateDto)
-  unitsSummary?: UnitsSummaryUpdateDto[]
-}
-
-// add other listing filter params here
-export class ListingFilterParams extends BaseFilter {
-  @Expose()
-  @ApiProperty({
-    type: String,
-    example: "Coliseum",
-    required: false,
-  })
-  @IsOptional({ groups: [ValidationsGroupsEnum.default] })
-  @IsString({ groups: [ValidationsGroupsEnum.default] })
-  [ListingFilterKeys.name]?: string;
-
-  @Expose()
-  @ApiProperty({
-    enum: Object.keys(ListingStatus),
-    example: "active",
-    required: false,
-  })
-  @IsOptional({ groups: [ValidationsGroupsEnum.default] })
-  @IsEnum(ListingStatus, { groups: [ValidationsGroupsEnum.default] })
-  [ListingFilterKeys.status]?: ListingStatus;
-
-  @Expose()
-  @ApiProperty({
-    type: String,
-    example: "Fox Creek",
-    required: false,
-  })
-  @IsOptional({ groups: [ValidationsGroupsEnum.default] })
-  @IsString({ groups: [ValidationsGroupsEnum.default] })
-  [ListingFilterKeys.neighborhood]?: string;
-
-  @Expose()
-  @ApiProperty({
-    type: Number,
-    example: "3",
-    required: false,
-  })
-  @IsOptional({ groups: [ValidationsGroupsEnum.default] })
-  @IsNumberString({}, { groups: [ValidationsGroupsEnum.default] })
-  [ListingFilterKeys.bedrooms]?: number;
-
-  @Expose()
-  @ApiProperty({
-    type: String,
-    example: "48211",
-    required: false,
-  })
-  @IsOptional({ groups: [ValidationsGroupsEnum.default] })
-  [ListingFilterKeys.zipcode]?: string;
-
-  @Expose()
-  @ApiProperty({
-    type: String,
-    example: "FAB1A3C6-965E-4054-9A48-A282E92E9426",
-    required: false,
-  })
-  @IsOptional({ groups: [ValidationsGroupsEnum.default] })
-  @IsString({ groups: [ValidationsGroupsEnum.default] })
-  [ListingFilterKeys.leasingAgents]?: string
-}
-
-export class ListingsQueryParams extends PaginationAllowsAllQueryParams {
-  @Expose()
-  @ApiProperty({
-    name: "filter",
-    required: false,
-    type: [String],
-    items: {
-      $ref: getSchemaPath(ListingFilterParams),
-    },
-    example: { $comparison: "=", status: "active", name: "Coliseum" },
-  })
-  @IsOptional({ groups: [ValidationsGroupsEnum.default] })
-  @IsArray({ groups: [ValidationsGroupsEnum.default] })
-  @ArrayMaxSize(16, { groups: [ValidationsGroupsEnum.default] })
-  @Type(() => ListingFilterParams)
-  @ValidateNested({ groups: [ValidationsGroupsEnum.default], each: true })
-  filter?: ListingFilterParams[]
-
-  @Expose()
-  @ApiProperty({
-    name: "view",
-    required: false,
-    type: String,
-  })
-  @IsOptional({ groups: [ValidationsGroupsEnum.default] })
-  @IsString({ groups: [ValidationsGroupsEnum.default] })
-  view?: string
-
-  @Expose()
-  @ApiProperty({
-    name: "orderBy",
-    required: false,
-    enum: OrderByFieldsEnum,
-    enumName: "OrderByFieldsEnum",
-    example: "updatedAt",
-  })
-  @IsOptional({ groups: [ValidationsGroupsEnum.default] })
-  @IsEnum(OrderByFieldsEnum, { groups: [ValidationsGroupsEnum.default] })
-  orderBy?: OrderByFieldsEnum
-
-  @Expose()
-  @ApiProperty({
-    type: String,
-    required: false,
-  })
-  @IsOptional({ groups: [ValidationsGroupsEnum.default] })
-  @IsString({ groups: [ValidationsGroupsEnum.default] })
-  jsonpath?: string
-}
-
-export class ListingsRetrieveQueryParams {
-  @Expose()
-  @ApiProperty({
-    name: "view",
-    required: false,
-    type: String,
-  })
-  @IsOptional({ groups: [ValidationsGroupsEnum.default] })
-  @IsString({ groups: [ValidationsGroupsEnum.default] })
-  view?: string
-}
-
-// Using a record lets us enforce that all types are handled in addFilter
-export const filterTypeToFieldMap: Record<keyof typeof ListingFilterKeys, string> = {
-  status: "listings.status",
-  name: "listings.name",
-  neighborhood: "property.neighborhood",
-  bedrooms: "unitTypeRef.num_bedrooms",
-  zipcode: "buildingAddress.zipCode",
-  leasingAgents: "leasingAgents.id",
-}
