+++ backend/core/src/listings/dto/listings-retrieve-query-params.ts
@@ -1,0 +1,16 @@
+import { Expose } from "class-transformer"
+import { ApiProperty } from "@nestjs/swagger"
+import { IsOptional, IsString } from "class-validator"
+import { ValidationsGroupsEnum } from "../../shared/types/validations-groups-enum"
+
+export class ListingsRetrieveQueryParams {
+  @Expose()
+  @ApiProperty({
+    name: "view",
+    required: false,
+    type: String,
+  })
+  @IsOptional({ groups: [ValidationsGroupsEnum.default] })
+  @IsString({ groups: [ValidationsGroupsEnum.default] })
+  view?: string
+}
