+++ backend/core/src/listings/dto/paginated-listing.dto.ts
@@ -1,0 +1,4 @@
+import { PaginationFactory } from "../../shared/dto/pagination.dto"
+import { ListingDto } from "./listing.dto"
+
+export class PaginatedListingDto extends PaginationFactory<ListingDto>(ListingDto) {}
