+++ backend/core/src/listings/listings.controller.ts
@@ -20,15 +20,7 @@ import {
 import { ListingsService } from "./listings.service"
 import { ApiBearerAuth, ApiExtraModels, ApiOperation, ApiTags } from "@nestjs/swagger"
 import { Cache } from "cache-manager"
-import {
-  ListingCreateDto,
-  ListingDto,
-  ListingUpdateDto,
-  PaginatedListingDto,
-  ListingsQueryParams,
-  ListingFilterParams,
-  ListingsRetrieveQueryParams,
-} from "./dto/listing.dto"
+import { ListingDto } from "./dto/listing.dto"
 import { ResourceType } from "../auth/decorators/resource-type.decorator"
 import { OptionalAuthGuard } from "../auth/guards/optional-auth.guard"
 import { AuthzGuard } from "../auth/guards/authz.guard"
@@ -36,6 +28,14 @@ import { mapTo } from "../shared/mapTo"
 import { defaultValidationPipeOptions } from "../shared/default-validation-pipe-options"
 import { Language } from "../shared/types/language-enum"
 import { ListingLangCacheInterceptor } from "../cache/listing-lang-cache.interceptor"
+import { PaginatedListingDto } from "./dto/paginated-listing.dto"
+import { ListingCreateDto } from "./dto/listing-create.dto"
+import { ListingUpdateDto } from "./dto/listing-update.dto"
+import { ListingFilterParams } from "./dto/listing-filter-params"
+import { ListingsQueryParams } from "./dto/listings-query-params"
+import { ListingsRetrieveQueryParams } from "./dto/listings-retrieve-query-params"
+import { ListingCreateValidationPipe } from "./validation-pipes/listing-create-validation-pipe"
+import { ListingUpdateValidationPipe } from "./validation-pipes/listing-update-validation-pipe"
 
 @Controller("listings")
 @ApiTags("listings")
@@ -43,7 +43,6 @@ import { ListingLangCacheInterceptor } from "../cache/listing-lang-cache.interce
 @ResourceType("listing")
 @ApiExtraModels(ListingFilterParams)
 @UseGuards(OptionalAuthGuard, AuthzGuard)
-@UsePipes(new ValidationPipe(defaultValidationPipeOptions))
 export class ListingsController {
   cacheKeys: string[]
   constructor(
@@ -57,12 +56,14 @@ export class ListingsController {
   @ApiOperation({ summary: "List listings", operationId: "list" })
   // ClassSerializerInterceptor has to come after CacheInterceptor
   @UseInterceptors(CacheInterceptor, ClassSerializerInterceptor)
+  @UsePipes(new ValidationPipe(defaultValidationPipeOptions))
   public async getAll(@Query() queryParams: ListingsQueryParams): Promise<PaginatedListingDto> {
     return mapTo(PaginatedListingDto, await this.listingsService.list(queryParams))
   }
 
   @Post()
   @ApiOperation({ summary: "Create listing", operationId: "create" })
+  @UsePipes(new ListingCreateValidationPipe(defaultValidationPipeOptions))
   async create(@Body() listingDto: ListingCreateDto): Promise<ListingDto> {
     const listing = await this.listingsService.create(listingDto)
     await this.cacheManager.reset()
@@ -72,6 +73,7 @@ export class ListingsController {
   @Get(`:listingId`)
   @ApiOperation({ summary: "Get listing by id", operationId: "retrieve" })
   @UseInterceptors(ListingLangCacheInterceptor, ClassSerializerInterceptor)
+  @UsePipes(new ValidationPipe(defaultValidationPipeOptions))
   async retrieve(
     @Headers("language") language: Language,
     @Param("listingId") listingId: string,
@@ -88,6 +90,7 @@ export class ListingsController {
 
   @Put(`:listingId`)
   @ApiOperation({ summary: "Update listing by id", operationId: "update" })
+  @UsePipes(new ListingUpdateValidationPipe(defaultValidationPipeOptions))
   async update(
     @Param("listingId") listingId: string,
     @Body() listingUpdateDto: ListingUpdateDto
@@ -99,6 +102,7 @@ export class ListingsController {
 
   @Delete(`:listingId`)
   @ApiOperation({ summary: "Delete listing by id", operationId: "delete" })
+  @UsePipes(new ValidationPipe(defaultValidationPipeOptions))
   async delete(@Param("listingId") listingId: string) {
     await this.listingsService.delete(listingId)
     await this.cacheManager.reset()
